INTEGRATION: Integration of 5035,5056,5068 from parent to child (branch spec: dev_to_lu_dev).
[git-p4: depot-paths = "//open/lu/dev/": change = 5074]

+++ farrago/unitsql/jdbc/metadata.sql
@@ -32,7 +32,7 @@ create schema metadata_test_schema;
 create table metadata_test_schema.new_table (
     tid integer primary key,
     address varchar(256),
-    amount integer );
+    amount decimal(5,2) );
 
 !tables
 !columns NEW_TABLE
